# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AgentOut": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "inflight": {
            "title": "Inflight",
            "type": "integer"
          },
          "max_concurrency": {
            "title": "Max Concurrency",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "stats": {
            "$ref": "#/components/schemas/AgentStatsOut"
          },
          "timeout_s": {
            "title": "Timeout S",
            "type": "number"
          }
        },
        "required": [
          "name",
          "display_name",
          "description",
          "model",
          "reasoning_effort",
          "timeout_s",
          "max_concurrency",
          "inflight",
          "fingerprint",
          "stats"
        ],
        "title": "AgentOut",
        "type": "object"
      },
      "AgentStatsOut": {
        "properties": {
          "executions": {
            "default": 0,
            "title": "Executions",
            "type": "integer"
          },
          "failed_executions": {
            "default": 0,
            "title": "Failed Executions",
            "type": "integer"
          },
          "labelled": {
            "default": 0,
            "title": "Labelled",
            "type": "integer"
          },
          "mean_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Cost Usd"
          },
          "p50_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P50 Latency Ms"
          },
          "p95_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Latency Ms"
          },
          "refreshed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refreshed At"
          },
          "success_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Rate"
          },
          "total_cost_usd": {
            "default": 0.0,
            "title": "Total Cost Usd",
            "type": "number"
          }
        },
        "title": "AgentStatsOut",
        "type": "object"
      },
...
      "MeOut": {
        "properties": {
          "api_key_id": {
            "format": "uuid",
            "title": "Api Key Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rate_limit_per_minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Per Minute"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "api_key_id",
          "name",
          "role",
          "rate_limit_per_minute"
        ],
        "title": "MeOut",
        "type": "object"
      },
...
      "ProblemOut": {
        "description": "RFC 9457 problem details, returned for every error.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Validation errors (422).",
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "instance",
          "request_id"
        ],
        "title": "ProblemOut",
        "type": "object"
      },
...
      "StrategiesOut": {
        "properties": {
          "adaptive_config": {
            "additionalProperties": true,
            "title": "Adaptive Config",
            "type": "object"
          },
          "strategies": {
            "items": {
              "$ref": "#/components/schemas/StrategyOut"
            },
            "title": "Strategies",
            "type": "array"
          }
        },
        "required": [
          "strategies",
          "adaptive_config"
        ],
        "title": "StrategiesOut",
        "type": "object"
      },
      "StrategyOut": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "available",
          "is_default",
          "description"
        ],
        "title": "StrategyOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nAdaptiveRoute routes each query to one of five specialised LLM agents\n(code, math, sql, writer, knowledge) using one of four strategies:\n**round_robin**, **embedding**, **llm** and **adaptive**.\n\nAuthenticate with `Authorization: Bearer <api key>` (or `X-API-Key`). Errors use\n`application/problem+json`. POST /v1/queries accepts an `Idempotency-Key` header.\n",
    "title": "AdaptiveRoute API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agents": {
      "get": {
        "operationId": "list_agents_v1_agents_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentOut"
                  },
                  "title": "Response List Agents V1 Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Validation error (or Idempotency-Key reused with a different body)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Rate limit exceeded (see Retry-After)"
          }
        },
        "summary": "List agents with live stats",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/me": {
      "get": {
        "description": "Lets clients adapt to the caller's role (e.g. hide admin-only actions).",
        "operationId": "whoami_v1_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Validation error (or Idempotency-Key reused with a different body)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Rate limit exceeded (see Retry-After)"
          }
        },
        "summary": "Who am I (the caller's API key)",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/strategies": {
      "get": {
        "operationId": "list_strategies_v1_strategies_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategiesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Validation error (or Idempotency-Key reused with a different body)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Rate limit exceeded (see Retry-After)"
          }
        },
        "summary": "List routing strategies",
        "tags": [
          "catalog"
        ]
      }
    },
...
  },
...
}